requirements.txt
Django==4.1.3
djangorestframework==3.14.0
django-rest-passwordreset==1.3.0
django-rest-knox==4.2.0
Pillow==9.5.0